+++ src/main/java/com/jelly/zzirit/domain/order/service/message/OrderConfirmConsumer.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import com.jelly.zzirit.domain.order.entity.Order;
 import com.jelly.zzirit.domain.order.repository.OrderRepository;
 import com.jelly.zzirit.domain.order.repository.PaymentRepository;
-import com.jelly.zzirit.domain.order.service.order.CommandConfirmService;
+import com.jelly.zzirit.domain.order.service.order.manage.CommandConfirmService;
 import com.jelly.zzirit.domain.order.service.pay.CommandRefundService;
 import com.jelly.zzirit.global.dto.BaseResponseStatus;
 import com.jelly.zzirit.global.exception.custom.InvalidOrderException;
