+++ src/main/java/com/jelly/zzirit/domain/order/service/order/cancel/OrderCancellationFacade.java
@@ -1,4 +1,4 @@
-package com.jelly.zzirit.domain.order.service.order;
+package com.jelly.zzirit.domain.order.service.order.cancel;
 
 import static com.jelly.zzirit.global.dto.BaseResponseStatus.*;
 
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import com.jelly.zzirit.domain.member.entity.Member;
 import com.jelly.zzirit.domain.order.entity.Order;
 import com.jelly.zzirit.domain.order.repository.OrderRepository;
+import com.jelly.zzirit.domain.order.service.order.CommandStockService;
 import com.jelly.zzirit.domain.order.service.pay.CommandRefundService;
 import com.jelly.zzirit.global.exception.custom.InvalidOrderException;
