+++ src/main/java/com/jelly/zzirit/domain/order/service/pay/CommandPaymentInitService.java
@@ -7,8 +7,8 @@ import com.jelly.zzirit.domain.member.repository.MemberRepository;
 import com.jelly.zzirit.domain.order.dto.request.PaymentRequest;
 import com.jelly.zzirit.domain.order.dto.response.PaymentInitResponse;
 import com.jelly.zzirit.domain.order.entity.Order;
-import com.jelly.zzirit.domain.order.service.order.CommandOrderSequence;
-import com.jelly.zzirit.domain.order.service.order.CommandTempOrderService;
+import com.jelly.zzirit.domain.order.service.order.manage.CommandOrderSequence;
+import com.jelly.zzirit.domain.order.service.order.manage.CommandTempOrderService;
 import com.jelly.zzirit.global.AuthMember;
 import com.jelly.zzirit.global.dto.BaseResponseStatus;
 import com.jelly.zzirit.global.exception.custom.InvalidUserException;
