+++ src/test/java/com/jelly/zzirit/domain/order/controller/order/OrderControllerRabbitTest.java
@@ -10,7 +10,6 @@ import static com.jelly.zzirit.domain.order.domain.fixture.OrderFixture.*;
 import static com.jelly.zzirit.domain.order.domain.fixture.OrderItemFixture.*;
 import static com.jelly.zzirit.global.dto.BaseResponseStatus.*;
 import static io.restassured.RestAssured.*;
-import static org.assertj.core.api.AssertionsForClassTypes.*;
 import static org.mockito.Mockito.*;
 import static org.springframework.restdocs.payload.JsonFieldType.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
@@ -43,7 +42,7 @@ import com.jelly.zzirit.domain.order.entity.OrderStatus;
 import com.jelly.zzirit.domain.order.entity.Payment;
 import com.jelly.zzirit.domain.order.repository.OrderRepository;
 import com.jelly.zzirit.domain.order.service.order.CommandStockService;
-import com.jelly.zzirit.domain.order.service.order.OrderCancelValidator;
+import com.jelly.zzirit.domain.order.service.order.cancel.OrderCancelValidator;
 import com.jelly.zzirit.domain.order.service.pay.CommandRefundService;
 import com.jelly.zzirit.domain.order.service.payment.TossPaymentClient;
 import com.jelly.zzirit.global.dto.BaseResponseStatus;
