+++ src/test/java/com/jelly/zzirit/domain/order/service/order/CommandOrderSequenceTest.java
@@ -3,7 +3,6 @@ package com.jelly.zzirit.domain.order.service.order;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
 
-import java.time.Duration;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
@@ -14,6 +13,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
+
+import com.jelly.zzirit.domain.order.service.order.manage.CommandOrderSequence;
 
 @ExtendWith(MockitoExtension.class)
 class CommandOrderSequenceTest {
